feat: refactored HostSidePanel to HostDrawer
