+++ apps/journeys-admin/src/components/Editor/Drawer/Attributes/blocks/Footer/Footer.spec.tsx
@@ -21,7 +21,7 @@ import {
 import { TestEditorState } from '../../../../../../libs/TestEditorState'
 
 import { Footer } from './Footer'
-import { HostSidePanel } from './HostSidePanel'
+import { HostDrawer } from './HostDrawer'
 
 jest.mock('react-i18next', () => ({
   __esModule: true,
@@ -150,7 +150,7 @@ describe('Footer', () => {
     expect(dispatch).toHaveBeenCalledWith({
       type: 'SetDrawerPropsAction',
       mobileOpen: true,
-      children: <HostSidePanel />
+      children: <HostDrawer />
     })
   })
