+++ apps/journeys-admin/src/components/Editor/Drawer/Attributes/blocks/Footer/HostDrawer/HostFormDrawer/index.ts
@@ -1,0 +1,1 @@
+export { HostFormDrawer } from './HostFormDrawer'
