+++ apps/journeys-admin/src/components/Editor/Drawer/Attributes/blocks/Footer/HostDrawer/HostInfoDrawer/index.ts
@@ -1,0 +1,1 @@
+export { HostInfoDrawer } from './HostInfoDrawer'
