+++ apps/journeys-admin/src/components/Editor/Drawer/Attributes/blocks/Footer/HostDrawer/HostListDrawer/index.ts
@@ -1,0 +1,1 @@
+export { HostListDrawer } from './HostListDrawer'
