+++ apps/journeys-admin/src/components/Editor/Drawer/Attributes/blocks/Footer/HostDrawer/index.ts
@@ -1,0 +1,1 @@
+export { HostDrawer } from './HostDrawer'
