--- apps/journeys-admin/src/components/Editor/Drawer/Attributes/blocks/Footer/HostSidePanel/HostForm/index.ts
@@ -1,1 +1,0 @@
-export { HostForm } from './HostForm'
