--- apps/journeys-admin/src/components/Editor/Drawer/Attributes/blocks/Footer/HostSidePanel/index.ts
@@ -1,1 +1,0 @@
-export { HostSidePanel } from './HostSidePanel'
